@@ -8,41 +8,44 @@
         "url": "https://github.com/Becklyn/kaba.git"
     },
     "version": "8.1.0",
-    "main": "lib/Kaba.js",
+    "main": "src/Kaba.js",
     "bin": {
         "kaba": "bin/run.js"
     },
     "scripts": {
-        "test": "ava"
+        "build": "node_modules/.bin/tsc --noEmitOnError --noErrorTruncation --listEmittedFiles --pretty --noUnusedLocals",
+        "dev": "node_modules/.bin/tsc --noEmitOnError --noErrorTruncation --listEmittedFiles --pretty",
+        "prepublishOnly": "npm run-script build",
+        "test": "npm run-script build && ava"
     },
     "dependencies": {
         "@babel/core": "^7.0.0",
         "@becklyn/typescript-error-formatter": "^1.0.4",
-        "babel-eslint": "^10.0.0",
+        "babel-eslint": "^10.0.2",
         "babel-loader": "^8.0.2",
-        "clean-webpack-plugin": "^3.0.0",
-        "eslint": "^5.5.0",
+        "core-js": "^3.1.4",
+        "eslint": "^6.0.1",
         "eslint-loader": "^2.1.0",
-        "eslint-plugin-react": "^7.11.0",
-        "fork-ts-checker-webpack-plugin": "^1.3.4",
+        "eslint-plugin-react": "^7.14.2",
         "fs-extra": "^8.0.0",
-        "kaba-babel-preset": "^2.0.3",
+        "kaba-babel-preset": "^4.0.1",
         "kleur": "^3.0.3",
         "pretty-hrtime": "^1.0.3",
-        "raw-loader": "^2.0.0",
-        "sade": "^1.4.1",
+        "raw-loader": "^3.0.0",
+        "sade": "^1.6.0",
         "terser-webpack-plugin": "^1.0.2",
-        "ts-loader": "^6.0.0",
-        "typescript": "^3.0.0",
-        "webpack": "^4.16.1"
+        "ts-loader": "^6.0.4",
+        "typescript": "^3.5.2",
+        "webpack": "^4.35.0"
     },
     "peerDependencies": {
         "kaba-scss": "^3.0.0",
         "webpack-bundle-analyzer": "^3.3.0"
     },
     "devDependencies": {
+        "@types/terser-webpack-plugin": "^1.2.1",
         "ava": "^2.0.0",
-        "kaba-scss": "^3.0.0",
+        "kaba-scss": "^3.0.1",
         "webpack-bundle-analyzer": "^3.3.2"
     },
     "engines": {
